Rename Hooks & Export Sign In & Up API Handlers
