+++ src/pages/Auth/Login/Login.tsx
@@ -7,8 +7,9 @@ import { TLoginFormInputs, loginSchema } from "../model";
 import TextField from "../../../components/common/TextField/TextField";
 import Button from "../../../components/common/Button/Button";
 import AuthLayout from "../AuthLayout";
-import { useAuth } from "../../../utils/hooks";
+import { useAuth } from "../../../hooks/useAuth";
 import { useMutation } from "@tanstack/react-query";
+import { handleLoginAPI } from "../../../api/auth";
 
 const Login: FunctionComponent = () => {
 	useAuth();
@@ -22,30 +23,8 @@ const Login: FunctionComponent = () => {
 		resolver: zodResolver(loginSchema),
 	});
 
-	const handleLogin = async (data: TLoginFormInputs) => {
-		try {
-			const url = import.meta.env.DEV
-				? import.meta.env.VITE_DEV_API
-				: import.meta.env.VITE_PROD_URL;
-
-			const result = await fetch(`${url}/api/users/auth/login`, {
-				method: "POST",
-				headers: { "Content-Type": "application/json" },
-				body: JSON.stringify(data),
-				credentials: "include",
-			});
-
-			const json = await result.json();
-			const { access_token } = await json;
-
-			localStorage.setItem("foliolinks_access_token", access_token);
-		} catch (error) {
-			console.log(error);
-		}
-	};
-
 	const loginMutation = useMutation({
-		mutationFn: handleLogin,
+		mutationFn: handleLoginAPI,
 		onSuccess: () => {
 			navigate("/dashboard");
 		},
