+++ src/pages/Auth/Register/Register.tsx
@@ -6,9 +6,10 @@ import { TRegisterFormInputs, registerSchema } from "../model";
 import TextField from "../../../components/common/TextField/TextField";
 import Button from "../../../components/common/Button/Button";
 import AuthLayout from "../AuthLayout";
-import { useAuth } from "../../../utils/hooks";
+import { useAuth } from "../../../hooks/useAuth";
 import { useMutation } from "@tanstack/react-query";
 import { useState } from "react";
+import { handleRegisterAPI } from "../../../api/auth";
 
 const Register = () => {
 	useAuth();
@@ -23,33 +24,8 @@ const Register = () => {
 
 	const [error, setError] = useState("");
 
-	const handleRegister = async (data: TRegisterFormInputs) => {
-		const url = import.meta.env.DEV
-			? import.meta.env.VITE_DEV_API
-			: import.meta.env.VITE_PROD_URL;
-
-		const result = await fetch(`${url}/api/users/auth/register`, {
-			method: "POST",
-			headers: { "Content-Type": "application/json" },
-			body: JSON.stringify(data),
-			credentials: "include",
-		});
-
-		const json = await result.json();
-
-		if (json.error) throw new Error(json.error);
-
-		const { access_token } = await json;
-
-		if (!access_token) throw new Error("Token not found");
-
-		localStorage.setItem("foliolinks_access_token", access_token);
-
-		return access_token;
-	};
-
 	const registerMutation = useMutation({
-		mutationFn: handleRegister,
+		mutationFn: handleRegisterAPI,
 		onSuccess: () => {
 			navigate("/dashboard");
 		},
